--- examples/IR_matplot_agent/control_param.py
@@ -1,4 +1,0 @@
-from cognify.hub.search import datavis
-
-# ================= Overall Control Parameter =================
-optimize_control_param = datavis.create_search(opt_log_dir='opt_results', evaluator_batch_size=50)
