--- examples/IR_matplot_agent/o1.py
@@ -1,17 +1,0 @@
-from openai import OpenAI
-client = OpenAI()
-
-response = client.chat.completions.create(
-  model="o1-mini",
-  messages=[
-    {
-      "role": "user",
-      "content": [
-        {
-          "type": "text",
-          "text": "Create a combination chart from the \\\"data.csv\\\" dataset, which contains sales data for various mobile phone brands across different quarters of the year. The columns in the CSV file are [\\\"Quarter\\\", \\\"Samsung\\\", \\\"Nokia/Microsoft\\\", \\\"Apple\\\", \\\"LG\\\", \\\"ZTE\\\", \\\"Huawei\\\"]. For each brand, create a box plot to represent the distribution of their sales data. On each box plot, display all the individual sales data points for that brand. Additionally, calculate the average sales for each brand and draw a line connecting these average values across the box plots. Use a consistent color scheme for the same quarter across different years, varying the shades to distinguish between years. Include a legend to aid in understanding the color coding."
-        }
-      ]
-    },
-  ]
-)
